Update XPCS GUI for easier multiplots
