+++ xicam/XPCS/__init__.py
@@ -17,18 +17,11 @@ from xicam.gui.widgets.imageviewmixins import PolygonROI
 from pyqtgraph.parametertree import ParameterTree, Parameter
 from pyqtgraph.parametertree.parameterTypes import ListParameter
 from .workflows import OneTime, TwoTime, FourierAutocorrelator
-from .widgets.onetimeplot import OneTimePlotWidget
+from .widgets.views import CorrelationView, FileSelectionView, OneTimeView, TwoTimeView
 
-# from . import CorrelationPlugin
 import time
 
-# class TwoTimeProcess(ProcessingPlugin):
-#     ...
-#
-#
-
 import numpy as np # TODO -- this is for debugging
-from .CorrelationDocument import CorrelationDocument
 import event_model
 from intake_bluesky.in_memory import BlueskyInMemoryCatalog
 from functools import partial
@@ -118,119 +111,13 @@ class OneTimeAlgorithms(ProcessingAlgorithms):
         return OneTime.name
 
 
-class CorrelationView(QWidget):
-    """
-    Widget for viewing the correlation results.
-
-    This could be generalized into a ComboPlotView / PlotView / ComboView ...
-    """
-    def __init__(self, model):
-        super(CorrelationView, self).__init__()
-        self.resultslist = QComboBox()
-
-        self.plot = pg.PlotWidget()
-
-        self.model = model # type: QStandardItemModel
-        self.resultslist.setModel(self.model)
-        self.selectionmodel = self.resultslist.view().selectionModel()
-
-        layout = QVBoxLayout()
-        layout.addWidget(self.resultslist)
-        layout.addWidget(self.plot)
-        self.setLayout(layout)
-
-        # Update plot figure whenever selected result changes
-        self.selectionmodel.selectionChanged.connect(self.updatePlot)
-
-    def results(self, selection: QItemSelection, dataKey):
-        results = []
-        for index in selection.indexes():
-            # for result in self.model.itemFromIndex(index).header['data'][dataKey]:
-            #     results.append(result)
-            results.append(self.model.data(index, Qt.UserRole)['data'][dataKey].value)
-            # results.append(self.model.itemFromIndex(index).data(DOC_ROLE)['data'][dataKey].value)
-        return results
-
-    def updatePlot(self, selected, deselected):
-        # TODO -- when multiple items selected to process, 2 items are added to the plot combo box (2 plot
-        # items exist in the plot widget, which is good though)
-        self.plot.clear()
-        for result in self.results(selected, 'g2'):
-            yData = result.squeeze()
-            # Offset x axis by 1 to avoid log10(0) runtime error at PlotDataItem.py:531
-            xData = np.arange(1, len(yData) + 1)
-            self.plot.plot(x=xData, y=yData)
-        self.resultslist.setCurrentIndex(0)#selected.indexes()[:-1].row())
-
-
-class OneTimeView(CorrelationView):
-    def __init__(self):
-        self.model = QStandardItemModel()
-        super(OneTimeView, self).__init__(self.model)
-        plotItem = self.plot.getPlotItem()
-        plotItem.setLabel('left', 'g<sub>2</sub>(&tau;)', 's')
-        plotItem.setLabel('bottom', '&tau;', 's')
-        plotItem.setLogMode(x=True)
-
-
-class TwoTimeView(CorrelationView):
-    def __init__(self):
-        self.model = QStandardItemModel()
-        super(TwoTimeView, self).__init__(self.model)
-        plotItem = self.plot.getPlotItem()
-        plotItem.setLabel('left', 't<sub>2</sub>', 's')
-        plotItem.setLabel('bottom', 't<sub>1</sub>', 's')
-
-
-class FileSelectionView(QWidget):
-    """
-    Widget for viewing and selecting the loaded files.
-    """
-    def __init__(self, headermodel, selectionmodel):
-        """
-
-        Parameters
-        ----------
-        headermodel
-            The model to use in the file list view
-        selectionmodel
-            The selection model to use in the file list view
-        """
-        super(FileSelectionView, self).__init__()
-        # self.parameters = ParameterTree()
-        self.filelistview = QListView()
-        self.correlationname = QLineEdit()
-        self.correlationname.setPlaceholderText('Name of result')
-
-        layout = QVBoxLayout()
-        layout.addWidget(self.filelistview)
-        layout.addWidget(self.correlationname)
-        self.setLayout(layout)
-
-        self.headermodel = headermodel
-        self.selectionmodel = selectionmodel
-        self.filelistview.setModel(headermodel)
-        self.filelistview.setSelectionModel(selectionmodel)
-        self.filelistview.setEditTriggers(QAbstractItemView.NoEditTriggers)
-        self.filelistview.setSelectionMode(QAbstractItemView.ExtendedSelection)
-
-        # Make sure when the tabview selection model changes, the file list
-        # current item updates
-        self.selectionmodel.currentChanged.connect(
-            lambda current, _:
-                self.filelistview.setCurrentIndex(current)
-        )
-
-        self.selectionmodel.currentChanged.connect(
-            lambda current, _:
-                self.correlationname.setPlaceholderText(current.data())
-        )
-
-
 class XPCS(GUIPlugin):
     name = 'XPCS'
 
     def __init__(self):
+
+        # TODO Move catalog outside of this method
+        self.catalog = BlueskyInMemoryCatalog()
 
         self.resultsmodel = QStandardItemModel()
 
@@ -241,7 +128,6 @@ class XPCS(GUIPlugin):
         # Widgets
         self.calibrationsettings = pluginmanager.getPluginByName('xicam.SAXS.calibration',
                                                                  'SettingsPlugin').plugin_object
-        self.plotwidget = OneTimePlotWidget()
         self.processor = XPCSProcessor()
 
         # Toolbar
@@ -295,6 +181,7 @@ class XPCS(GUIPlugin):
         item = QStandardItem(header.startdoc.get('sample_name', '????'))
         item.header = header
         self.headermodel.appendRow(item)
+        self.selectionmodel.reset()
         self.selectionmodel.setCurrentIndex(self.headermodel.index(self.headermodel.rowCount() - 1, 0),
                                             QItemSelectionModel.Rows)
         self.headermodel.dataChanged.emit(QModelIndex(), QModelIndex())
@@ -373,24 +260,22 @@ class XPCS(GUIPlugin):
     def createDocument(self, view: CorrelationView=None):
         # self.correlationdocument.createEvent(name=data['name'], image_series=data['name'],
         #                                      g2=data['result']['g2'].value)
-        # TODO Move catalog outside of this method
-        catalog = BlueskyInMemoryCatalog()
-        catalog.upsert(self._createDocument, (self.__results,), {})
-        key = list(catalog)[0]
+        self.catalog.upsert(self._createDocument, (self.__results,), {})
+        key = list(self.catalog)[-1]
 
-        for name, doc in catalog[key].read_canonical():
+        for name, doc in self.catalog[key].read_canonical():
             if name == 'event':
                 resultsmodel = view.model
                 item = QStandardItem(doc['data']['name'])  # TODO -- make sure passed data['name'] is unique in model -> CHECK HERE
                 item.setData(doc, Qt.UserRole)
                 resultsmodel.appendRow(item)
                 selectionModel = view.selectionmodel
-                deselected = selectionModel.selection()
+                selectionModel.reset()
                 selectionModel.setCurrentIndex(
                     resultsmodel.index(resultsmodel.rowCount() - 1, 0), QItemSelectionModel.Rows)
                 selectionModel.select(selectionModel.currentIndex(), QItemSelectionModel.SelectCurrent)
                 # self.currentSelectionModel().selectionChanged.emit(self.currentSelectionModel().selection(), deselected)
-                resultsmodel.dataChanged.emit(QModelIndex(), QModelIndex())
+                # resultsmodel.dataChanged.emit(resultsmodel.currentIndex())
 
         self.__results = []
         print()
@@ -403,7 +288,8 @@ class XPCS(GUIPlugin):
         yield 'start', run_bundle.start_doc
 
         data_keys = {'image_series': {'source': 'Xi-cam', 'dtype': 'string', 'shape': []},
-                     'g2': {'source': 'Xi-cam XPCS', 'dtype': 'number', 'shape': [],},
+                     'g2': {'source': 'Xi-cam XPCS', 'dtype': 'number', 'shape': []},
+                     'lag_steps': {'source': 'Xi-cam XPCS', 'dtype': 'number', 'shape': []},
                      'name': {'source': 'Xi-cam', 'dtype': 'string', 'shape': []}
                      }
         stream_bundle = run_bundle.compose_descriptor(data_keys=data_keys, name='primary')
@@ -411,10 +297,12 @@ class XPCS(GUIPlugin):
 
         for result in results:
             yield 'event', stream_bundle.compose_event(data={'image_series': result['name'],
-                                                             'g2': result['result']['g2'],
+                                                             'g2': result['result']['g2'].value,
+                                                             'lag_steps': result['result']['lag_steps'].value,
                                                              'name': result['name']},
                                                        timestamps={'image_series': timestamp,
                                                                    'g2': timestamp,
+                                                                   'lag_steps': timestamp,
                                                                    'name': timestamp})
 
         yield 'stop', run_bundle.compose_stop()
